+++ R/check_quarto.R
@@ -1,62 +1,51 @@
-#' R6 classes
-#'
-#' Check classes
-#' @export
-check_quarto_pdf = R6::R6Class(
-  "check_quarto_pdf",
-  inherit = base_check,
-  public = list(
-    #' @description PDF rendering
-    check = function() {
-      private$checker(render_quarto(private$short))
-      return(invisible(NULL))
-    }
-  ),
-  private = list(
-    context = "Rendering",
-    short = "pdf",
-    group = "quarto"
+#' @rawNamespace export(check_quarto_beamer)
+#' @rawNamespace export(check_quarto_docx)
+#' @rawNamespace export(check_quarto_html)
+#' @rawNamespace export(check_quarto_observable)
+#' @rawNamespace export(check_quarto_pdf)
+NULL
+types = c("beamer", "docx", "html", "observable", "pdf")
+for (type in types) {
+
+  assign(
+    paste0("check_quarto_", type),
+    R6::R6Class(
+      paste0("check_quarto_", type),
+      inherit = uatBase::base_check,
+      public = list(
+        check = function(debug_level) {
+          quarto_dir = system.file("extdata", private$group, private$short,
+                                   package = "uatBase", mustWork = TRUE)
+          private$checker(
+            render_quarto(quarto_dir, debug_level = debug_level))
+
+          return(invisible(NULL))
+        }
+      ),
+      private = list(
+        context = paste("Rendering", type),
+        short = type,
+        group = "render_quarto"
+      )
+    )
   )
-)
+}
 
-#' R6 classes
-#'
-#' Check classes
-#' @export
-check_quarto_html = R6::R6Class(
-  "check_quarto_html",
-  inherit = base_check,
-  public = list(
-    #' @description HTML rendering
-    check = function() {
-      private$checker(render_quarto(private$short))
-      return(invisible(NULL))
-    }
-  ),
-  private = list(
-    context = "Rendering",
-    short = "html",
-    group = "quarto"
-  )
-)
+render_quarto = function(quarto_dir, debug_level) {
+  suppress = uatBase::get_suppress(debug_level) #nolint
 
-#' Test rendering
-#'
-#' Simple test for rendering. In theory, the quarto
-#' @param output_format File extension specifies the type to produce
-render_quarto = function(output_format = c("pdf", "html")) {
+  tmp_dir = file.path(tempdir(), "quarto")
+  on.exit({if (fs::dir_exists(tmp_dir)) fs::dir_delete(tmp_dir)}) #nolint
+  dir.create(tmp_dir, showWarnings = FALSE)
+  lapply(list.files(quarto_dir, full.names = TRUE),
+         function(f) file.copy(f, tmp_dir))
 
-  output_format = match.arg(output_format)
-  qmd_filepath = system.file("extdata", "test-quarto-document.qmd",
-                             package = "jrHealthCheckWorkbench", mustWork = TRUE)
-
-  output_file = paste0("uat-notebook.", output_format)
-
+  qmd_filepath = file.path(tmp_dir, "index.qmd")
   quarto::quarto_render(input = qmd_filepath,
-                        output_format = output_format,
-                        output_file = output_file, # Can't be a path
-                        quiet = TRUE)
-  doc = normalizePath(output_file)
-  if (!(file.info(doc)$size > 100)) stop("Quarto file too small")
+                        execute_dir = tmp_dir,
+                        quiet = (debug_level == 0),
+                        cache = FALSE,
+                        as_job = FALSE
+  )
   return(invisible(TRUE))
 }